requirements.txt
PyQt6>=6.6
PyQt6-Fluent-Widgets>=1.5
PyQt6-Frameless-Window>=0.4
APScheduler>=3.10
PyYAML>=6.0
psutil>=5.9
pywin32>=306
darkdetect>=0.8
